refactor(workbench): 提取桌面与 Web 共享运行时
将终端 Pane、运行时绑定与租约、拖拽状态机和窗口布局迁入 @liveagent/ui，桌面端保留宿主注入与兼容导出。

统一侧栏、Right Dock 和 SSH Overlay 的拖拽起点，并修复多 Pane 聚焦时的 composer 引用与草稿状态切换。
