+++ crates/agent-gui/src/pages/chat/surfaces/ConversationPaneHost.tsx
@@ -118,6 +118,8 @@ export const ConversationPaneHost = forwardRef<
   } = useConversationPaneBinding({ paneId, conversationId, project });
   const composerRef = useRef<MentionComposerHandle | null>(null);
   const scrollFollowRef = useRef<ScrollFollowHandle | null>(null);
+  const controllerRef = useRef(controller);
+  controllerRef.current = controller;
   const [composerOverlayHeight, setComposerOverlayHeight] = useState(0);
   // Background panes send through their own conversation-scoped pipeline; the
   // handler owns clear-on-send/restore-on-failure for this pane's composer.
@@ -144,9 +146,10 @@ export const ConversationPaneHost = forwardRef<
     [],
   );
 
+  // biome-ignore lint/correctness/useExhaustiveDependencies: keyed on conversationId so primary↔background controller swaps for the same conversation never clear a composer mid-typing.
   useLayoutEffect(() => {
     const composer = composerRef.current;
-    const draft = controller.getSnapshot().draft;
+    const draft = controllerRef.current.getSnapshot().draft;
     if (draft) {
       composer?.setDraft(draft);
     } else {
@@ -158,13 +161,16 @@ export const ConversationPaneHost = forwardRef<
       // cleared this composer mid-switch (legacy reset semantics), so an
       // empty composer must not delete the draft cached in the registry —
       // deliberate clears propagate through the page-level draft cache.
+      // Depend on conversationId, not controller: primary↔background binding
+      // swaps the controller object for the same conversation and must not
+      // clear a composer the user is mid-typing in.
       const nextDraft = composer?.getDraft();
       if (!nextDraft || nextDraft.isEmpty || !nextDraft.text.trim()) {
         return;
       }
-      controller.setDraft(nextDraft);
+      controllerRef.current.setDraft(nextDraft);
     };
-  }, [controller]);
+  }, [conversationId]);
 
   return (
     <ConversationSurface
