+++ crates/agent-ui/src/components/chat/ChatHistorySidebarTypes.ts
@@ -95,14 +95,24 @@ export type ChatHistorySidebarProps = {
   /** Workbench drag intent from a conversation row title (desktop pointer). */
   onConversationWorkbenchDragIntent?: (
     item: SidebarConversation,
-    event: { pointerId: number; clientX: number; clientY: number },
+    event: {
+      pointerId: number;
+      clientX: number;
+      clientY: number;
+      currentTarget?: EventTarget | null;
+    },
   ) => void;
   /** Menu alternative to dragging: open a conversation in a split pane. */
   onConversationOpenInWorkbenchSplit?: (item: SidebarConversation) => void;
   /** Workbench drag intent from a project row title (creates a conversation). */
   onProjectWorkbenchDragIntent?: (
     project: WorkspaceProject,
-    event: { pointerId: number; clientX: number; clientY: number },
+    event: {
+      pointerId: number;
+      clientX: number;
+      clientY: number;
+      currentTarget?: EventTarget | null;
+    },
   ) => void;
   onStartRenaming: (item: SidebarConversation) => void;
   onRenameDraftChange: (value: string) => void;
