+++ crates/agent-ui/src/components/project-tools/RightDockTabStrip.tsx
@@ -1,5 +1,5 @@
 import type { RightDockTabKind } from "@liveagent/app/lib/settings";
-import { Check, Columns2, Cpu, Terminal, X } from "@liveagent/ui/components/IconSet";
+import { Check, Columns2, Cpu, GripVertical, Terminal, X } from "@liveagent/ui/components/IconSet";
 import { useLocale } from "@liveagent/ui/i18n/index";
 import type { CSSProperties, ReactNode } from "react";
 import { useState } from "react";
@@ -31,14 +31,19 @@ type RightDockTabStripProps = {
   onCloseTerminalRequest: (session: TerminalSession) => void;
   /**
    * Provided when terminal tabs can be dragged out of the dock (workbench
-   * hosts). The tab body then arms the drag-out gesture instead of tab
-   * reorder; reorder stays available from the grip handle. Click activation
-   * is unaffected — the drag session suppresses the click only after its
-   * movement threshold.
+   * hosts). The grip and the tab body both arm the drag-out gesture; in-dock
+   * reorder yields because the visible handle is the extract affordance.
+   * Click activation is unaffected — the drag session suppresses the click
+   * only after its movement threshold.
    */
   onTerminalTabDragStart?: (
     session: TerminalSession,
-    event: { pointerId: number; clientX: number; clientY: number },
+    event: {
+      pointerId: number;
+      clientX: number;
+      clientY: number;
+      currentTarget?: EventTarget | null;
+    },
   ) => void;
   /**
    * Keyboard/pointer alternative to dragging a terminal tab out: docks the
@@ -151,7 +156,29 @@ export function RightDockTabStrip(props: RightDockTabStripProps) {
               : undefined
           }
         />
-        {renderTabDragHandle(tab.id, tab.label)}
+        {tab.dragProps ? (
+          <button
+            type="button"
+            data-project-tools-tab-action="drag"
+            aria-label={t("workbench.dragPane")}
+            title={t("workbench.dragPane")}
+            className={cn(
+              "relative z-10 flex h-6 w-5 shrink-0 items-center justify-center rounded text-muted-foreground/45 opacity-70 transition-[background-color,color,opacity] focus-visible:outline-none focus-visible:ring-1 focus-visible:ring-ring",
+              "cursor-grab touch-none hover:bg-background/80 hover:text-foreground hover:opacity-100 focus-visible:bg-background focus-visible:text-foreground focus-visible:opacity-100 active:cursor-grabbing",
+            )}
+            onPointerDown={(event) => {
+              // The reorder handle sits above the tab body and used to
+              // stopPropagation into beginTabDrag, so grabbing the only
+              // visible grip never extracted the session onto the canvas.
+              event.stopPropagation();
+              tab.dragProps?.onPointerDown(event);
+            }}
+          >
+            <GripVertical className="h-3.5 w-3.5" />
+          </button>
+        ) : (
+          renderTabDragHandle(tab.id, tab.label)
+        )}
         <div
           aria-hidden="true"
           className="pointer-events-none relative z-10 flex h-full min-w-0 flex-1 items-center gap-1.5 text-left text-inherit"
@@ -291,6 +318,7 @@ export function RightDockTabStrip(props: RightDockTabStripProps) {
                     pointerId: event.pointerId,
                     clientX: event.clientX,
                     clientY: event.clientY,
+                    currentTarget: event.currentTarget,
                   });
                 },
               }
